+++ NEWS.md
@@ -1,3 +1,9 @@
+develop
+==============
+
+1. make `maxResets` a function parameter
+2. provide optional (`optimalWarmStart = FALSE`) more parallelization when profiling likelihood
+
 Cyclops v3.4.1
 ==============